requirements.txt
ultralytics==8.2.62
opencv-python>=4.10.0
numpy>=1.26
pyyaml>=6.0.1
streamlit>=1.32.0
pdfminer.six>=20221105
